Changes to Motif Atlas output files Copies the VARNA files to the appropriate website directory.
N.B. Beware of the filesystem structure being different on test
and production.  Not a problem this time.

Copies and renames the mat output files and stores them in a
directory parallel to the VARNA files.  Ideally, this would be
contained in its own class, but it is logically quite similar to
the VARNA renaming process.