Test using "magic values" for pairs with bad alignments Currently, the pipeline is repeating calculations for groups that contain one or more bad alignments ad nauseum, and preventing the members of those groups that do have good alignments from having discrepancy data available to view.
First-pass attempt to resolve this by setting the discrepancy to
-1 and the number of pairs considered to 0.  Both of these values
are clearly invalid within the context of the data, and can be
readily excluded by code which evaluates discrepancy data.

TO DO:  develop logic that will enable easy recalculations for
this data.  Preferably at the pairs alignment level, though at
a correspondence_id would probably be acceptable in most cases.